added experimental support for nested pythonic attribute access
* supports `Node.a.very.long.attribute.path.expression`
* can be mixed with item access: `Node.a[("very", "long")].attribute["path"].expression`

warning: this is a bit dangerous due to the use of __getattr__.
SQLAlchemy uses that method sometimes to test for optional members.
In order to fix this, we ignore requests for attributes with \_underscores
and the special values `key` and `is_literal`.
As a consequence, attributes starting with underscores or called `key` or `is_literal` cannot be accessed this way.

Changes in SQLAlchemy may cause additional failures. If SQLAlchemy
removes the use of getattr, we may drop these special attribute name restrictions