correct child count handling
The old default implementation of childcount() in Data as "return 0"
doesn't make any sense. The default is now to return the count of all
content children in all subcontainers like it was done in
web.frontend.frame before for nav tree labels of directories.
Subclasses of Directory can override childcount() to define another method of "counting children".