web/frontend/streams.py get_archived: postgres, new interface
This commit replaces core.tree with db.query (as usual).
It also changes the way archive managers are told to fetch files.

The old archive manager interface "called" the archive
manager python module file in a new python interpreter,
then used the database to communicate with the process.
This concept seems to be fragile as
it relies on instant database updates,
irrespective of request boundaries.

The new interface simply invokes the
getArchivedFile method of the archive manager.
If there is a reason to start a separate process,
the archive manage has to do so by itself.