pdf files with drm cannot be uploaded, closes #1184
contenttypes/document.py: don't throw an exception for pdfs with drm

event_files_changed(): the upload of pdfs with drm throws an exception,
but this is not reported to the user. The exchange/add digital object
for pdfs with drm throws also an exception, but this is reported to the
user, so this must be inhibited.

contenttypes/document.py: enable upload of pdf files with drm

event_files_changed(): call db.session.commit() before raising the error
OperationException if document is encrypted to accept document