admin aerea refuses to create metafields with the same name, but cheating is possible, closes #1422
If one created (for example) a metafield `author` for a given schema, the admin area will refuse to create another metafield for this schema with the same name.
It is, however,  possible to create one with another name like `author2` and later change that name from `author2` to `author`.
Unfortunately this will lead to exceptions in several situations (view linked masks, deleting one of those two metafields, ...), because the names of the metafields are taken to be unique.

It should therefore be prevented to rename a metafield to a name already used.