Editor: remove start page "preview" function for containers, rename it for content nodes, closes #1497
For containers, the "preview" editor tab displays the current start page with wrong CSS. Remove this tab because isn't very helpful. The user can visit the frontend instead to see it in proper form.

For content nodes, the "preview" should be renamed to "display" because it shows the actual content, not a preview. We cannot remove it currently because it is the default tab and removing it could confuse users.