BibTeX Import: field "www-address" incorrect, closes #1513
The field "www-address" is separated by a ';' in a link and a link-text. Also if link-text is empty the ';' is needed at the end of the link otherwise the fields for link and link-text is empty if edit metadata is called.
So if link-text is empty no ';' should be needed at the end of the link