metadata schema: intermediate db commits slows down import, closes #1550
During an import of a metadata schema there are a lot of db commits which slows down the import dramatically.
Additionally a metadata schema with the same name is created during the import which leads to an exception if this metadata schema is edited during the import. This second metadata schema with the same name lasts as long as the import goes on. Since this import has finished the medata schema is renamed.