upload field: fix upload, limit to admins, closes #1566
Fix file upload because it is needed for the configuration of workflow steps of the type add PDF pages. Due to insufficient permission handling in the implementation, the uploading / deleting and listing of files is only allowed for admins. This means that the upload field type cannot be used by editors at the moment. It would be better to remove this feature or find a better solution.