Selecting a large collection in editor leads to slow page loads, closes #1017
All items in the collection are displayed in the editor after a long waiting time or a timeout is raised.

Solution could be to introduce pagination.
Selecting of all entries could be handled analogous to GoogleMail:
There, selecting all items on one pages makes a context menu appear that allows the selection of all items in this view.

Additional feature might be auto-loading of next page on scroll-down.

Do we want a search-feature in the edit-area?