pdf-upload: ignore utf-16 strings, closes #1611
During a pdf upload 'pdfinfo -meta <pdf-file>' is called to extract technical metadata and write them in an info-file. However pdfinfo cannot extract metadata which are coded in utf-16 format and clips them after BOM-UTF16-BE (Byte Order Mark for UTF-16). These strings must be ignored because they lead to problems if the info-file is later read again.