pdf files cannot be uploaded for nodes of type dissertation, closes #1672
With ticket 1445 the filetype of an uploaded file is checked against the node type. This leads to problems if a pdf file, which is from type 'document' should be uploaded to a node of type defined in a plugin. In this case also the method get_upload_filetype() must be used to check the correct filetype.