access to lower versions returns an error, closes #1683
The access to lower versions of an document returns an error. The reason is a new version of the package 'werkzeug'. In the new version cached_property is a subclass of property while in the old version cached_property a subclass of object. This leads to problems in the classmethod __getattr__ of the class MtVersionBase where first is checked if a type is instance of property and then if it is a instance of cached_property. So the sequence of tests must be changed and it must be test first of a type is instance of cached_property and then if is instance of property.