workflow diss: wrong mailaddress blocks workflow, closes #1764
If during workflow execution the sending of an email fails this is not reported. Not till then the node is called in the workflow again the correct reason of the failing email is reported. If the user then uses the link to resend the mail and the sending of the email fails again a second workflow_lock is acquired which blocks this workflow and blocks all other users which trying to use the workflow.