acceleration of simple search, closes #1701
With the correction for ticket #1413 a distinct statement is added to the psql search query to avoid double results. But this increases the execution time dramatically because the whole searchresult must be fetched by postgres to eliminate double entries, independently of a limitation statement. Additionally the psql search query is called three times per search command which leads to a further delay.
The reason for double entries in the search result are more then one configured search languages, e.g. german and english. In the most cases every configured search language produces the same hit for a query, which leads to double results. Because the double results are in a sequence, postgres test first one language then the other, double results are filtered by sqlalchemy and so the distinct statement is not necessary.