adding node as childnode to itself must be prevented, closes #1793
Normally a node cannot be selected as childnode for itself. But if the directory was changed and the previous directory is selected again, to own node can be selected as childnode. This must be prevented because the postgres transaction is blocked and only a restart of postgres clears this situation.