bibtex import: datefield must not be longer than format, closes #1791
If the bibtex import contains a datefield, this is in the most cases 'year', then the corresponding metadata must not have more character than in the format is specified, except the format contains '%'. This means if the datefield is year and the format is 'yyyy' then the result must not be longer than 4. E.g. if year is 2017 and format is 'yyyy' then the metadata year must be 2017 and not 2017-00-00T00:00:00.