node with childnodes: use same link for thumbnail and details, closes #1884
If in editmode 'details...' is clicked for a childnode (details) the childnode is displayed in the editor as it is expected.
But if the thumbnail is clicked the childnode is displayed in a frontend frame within the editor which is wrong.
So the link behind the thumbnail must be the same as for 'details...'