workflow: prevent creation/changing of a workflowstep with the name of an existing workflowstep, closes #1893
To prevent multiple workflowsteps with the same name in the same workflow, the creation of a new workflowstep with the name of an existing workflowstep must be inhibited. Also the changing of the name of an existing workflowstep to a name of another workflowstep existing already in the workflow must be inhibited.