Temporary files created with tempfile.mkstemp must be closed correctly, closes #1908
Temporary files, which are used for printing, are created with tempfile.mkstemp. But this function implicitly also opens the file. So this file must be closed when the file is no longer needed. Otherwise every print order increases the number of open file descriptors with the result, that the program is no longer functional if too many files are opened.