New configuration parameter for ldap-users: auto_create_user, closes #1920
The ldap section of mediatum.cfg gets a new entry: auto_create_user.
With this entry it can be controlled if a ldap-user is accepted if he knows the correct password for his ldap-account but has no entry in the user table.
auto_create_user may have the one of the following values:
-   always: the ldap-user is always accepted
-   group: the ldap-user is only accepted if he belongs to a ldap-group which is although configured in mediatum
-   never: the ldap-user is not accepted