Deny access to systemfiles with item parameter, closes #1941
It is possible to show the content of systemfiles with a request like GET /?item=../../../../../../../../../../etc/passwd.
This is a security violation and must be inhibited.
So only filenames must be allowed which starts with a number and the number is a node id and the current user has read access to this node.