navigationtree without cache, closes #2030
Because the database query to determine the navigation for guest users lasts several seconds (about 3 seconds with a node table of about 500000 entries) a cache was used for this database query only for guest users. This cache values are valid for 5 minutes and afterwards they are computed again if an guest user opens the navigation tree again.
But this is an security issue because if some collections or directories may be restricted only for guest users of an internal network.
If a guest user of an internal network opens a navigation tree with restricted collections or directories and this navigation tree is cached also guest users outside the internal network may see this restricted collections or directories for the next 5 minutes.
So the cache must be removed and the database query must speed up.