bibtex/doi-import: set all metadata, closes #2032
Documents import via bibtex or doi mostly have only a subset of metadata of the assigned schema.
All missing metadata in the document for this schema should now be set like it was made as if the document is loaded in editor with 'edit metadata' and saved without any changes.