metafields of type list does not show multiple selections, closes #2037
metafields of type list have a checkbox for multiple selections. If this checkbox is clicked in the metafield editor and the metafield is saved, then it is saved correctly in database with the attrs->>'multiple'='true'. But if the metafield is edited again the checkbox is not shown as marked although it marked in the database. So this in only a rendering problem.