ckeditor: utf8 character problems, closes #2039
The ckeditor which is used to edit startpages cannot handle special utf8 characters like ● (Black Circle). These characters are not translated in the correct html entity like &#9679; and mediatum gets the exception UnicodeEncodeError if the startpage is written back. The result is an empty startpage.
So mediatum should handle this exception and should try to translate this characters itself.