session expiration time, closes #2127
For every request without a PSESSION cookie a new session is created, but they were not deleted.
So the number of sessions may be growing endless during the live time of mediatum. This also means that if a user has logged in, he keeps logged in during the life time of mediatum until he logged off.
So a mechanism should be implemented that older sessions are deleted depending on their last use. Which also means that a user is automatically logged off, if he makes no action for specified time.
For this purpose a session expiration time should be defined which may be configurable.

If a request with a PSESSION cookie comes and the session can be found in the stored sessions, it should be first checked if the last use is older than the session expiration time. If so the session should be deleted and recreated again.
If a request comes without the PSESSION cookie and new session is created all previously stored sessions should be checked if their last use is older than the session expiration time and these sessions should be deleted. To avoid too many checks for older sessions, this should be only done if the last session check is a special time ago. This special time should be 1/10 of the session expiration time.