Acceleration of print function, closes #2141
Printing of directories with many nodes is very slow because for every node the "printlist" or "nodesmall" mask must be detected depending on the metadatatype of the node. Therefore use a maskcache with the metadatatype as key so the "printlist" or "nodesmall" mask must only be detected if the metadatatype is not in the maskcache.
The maskcache is only valid within the printfunction and is recreated if the printfunction is called again.