memory leak: sql statements not freed, closes #2120
mediatum uses sqlalchemy events to detect slow queries. Before a query the sql statement is saved and after the query the sql statement is reported and freed if the query duration lasts too long. The problem is that the sql statement is not freed if the query duration is within the time limit. So the list of sql statements is growing.
The sql statement must be also freed if the query is within the time limit.