print function not working, closes #2221
With ticket #2105 x-accel-redirect is not set for temporary files, because temporary files may be deleted before nginx can send them to the user.
With ticket #2141 the printfunction is accelerated and sendfile function is used where x-accel-redirect is set.
The effect is that the created printpdf is served by nginx, but if nginx runs not as root or not as mediatum, nginx has no read access to the printpdf and reports 403 Forbidden.
So x-accel-redirect must not be set for printpdf.