fts: create new primary key, closes #2173
The current primary key id for the table fts makes an update of the table node very slow because the trigger update_node_tsvector must do an seqential scan of the table fts to find node id nid for the node which is updated.
So a better primary key is the combination nid,config,searchtype.
For changing the primary key a new alembic version must be created.