workflows: read constraints remains after publishing after migration to postgres 10.3, closes #2329
The workflowstep publish removes read constraints from a document to allow access to the document for everyone, except there is a time limitation.
After the migration to postgres 10.3 the readconstraints are no longer removed in the workflowstep publish,
because the read accessrule of the document is compared with the first accessrule for the usergroup _workflow before it is removed.
The problem is, that the sequence of table access_rule is inverted after the migration to postgres 10.3. So the first accessrule does no longer match with the accessrule of the document, if the document is created in the workflow before the migration to postgres 10.3.
To solve this problem all accessrules belonging to usergroup _workflow must be compared, not only the first one.