Patch WTForms 2.2.1 to restore compatibility with SQLAlchemy 1.2+, closes #2343
WTForms calls a function from SQLAlchemy (if present).
But the return value of that function changed
with SQLAlchemy 1.2; WTForms gets more
than it asks for and an exception occurs.

Sadly, as of this moment, there is no new release
of WTForms that fixes the error, but it is fixed
in the current master branch in Github.
We import the patch from Github and
inject it into WTForms build script.