Cross-Site-Request-Forgery In Forms, closes #2021
The goal is to integrate a hidden field with an anti-CSRF token into every form on all websites. For this purpose wtforms is used. This form is instantiated for all forms at a central location in Athana. The anti-CSRF token of this form is then inserted in all views (j2.jade, html).