Athana: extract reuseable elements (#2447)
Athana generally consists of request processing part and
communication part based on client socket server asyncore.
The first part can be considered as the controller of
Model–View–Controller architectural pattern. The controller
accepts input and converts it to commands for the model or view.
Extract this controller to build an interface to any other web
framework, so that athana is still runnable. continue_request
function changed to handle_request is the entry point.
Later commits are web framework specific.