editor: show error message if create object fails (#2451)
If after an upload the button create objects is pressed,
the user gets no hint if create objects failed.
This should be changed in this way,
that the uploaded file is displayed in red and an
error message is displayed if create objects failed.
Additionally the upload popup window should not be closed
automatically, so the user has enough time to see the error
message and then close the upload popup window manually.