Secret key: flask unique secret_key (#2056)
According to flask documentation, the secret_key on the application
must be set to something unique and secret. Use uwsgi cache to get
key randomly generated but unique for multiple processes. The uwsgi
cache_set(key, value) function did not change the value of key if
it already is set.