utils: check around urlsplit for bad unicode (#2523)
`urlparse.urlsplit` in Python up to 2.7.16 fails to
handle some unicode characters correctly:
Unicode characters that resolve to URL-separating
characters like "@" aren't recognized and
could lead to unexpected behaviour of the
code processing the output of `urlsplit`.
If the input data is user-controlled,
this can be exploited by malicious users.

The problem is handled upstream in
https://bugs.python.org/issue36216
and resolved in Python 2.7.17 .

Due to technical reasons, we have to
postpone an update of Python 2.7.

This commit adds code to the entry point of the
main mediatum script that replaces the `urllib.urlsplit`
method by an implementation that checks
the output for bad unicode characters.
This should not only fix our uses of the function
but also third-party modules that use it.

Note that the code violates our code style recommendations,
mainly in order to patch `urllib.urlsplit` before
any other modules might get a chance to load `urllib`.