doi,bibtex import: expand datefields to standard format for publishing
if a doi or bibtex is imported and the metadata are edited without
saving, the date fields are not in the same format as if the edited
metadata are saved. This leads to problems if the data are published. So
the data fields are already formatted if doi or bibtex import takes
place. (#2591)