z39.50: fix some import errors
The flask transition preparations
yielded a new module `athana_z3950.py`.
However, some imports were missed while
adapting the code to the new file.

In particular, it turned out that simply
using `core.medmarc` in the Z39.50 handler
doesn't work as `medmarc` is not loaded.
It also cannot be loaded PEP8 compliant
at the module header as it must be loaded
after or at the end of `full_init`.
This kind of deferred module loding is frowned upon,
but `basic_init` already defer-loads other module
(probably due to the same reasons),
so we add `medmarc` to this list of module.