editor: truncated text may lead to invalid html tags
Dictionaries in editor displays the single nodes with the nodesmall mask in one line.
To avoid too long lines the result of the nodesmall mask is truncated after the 200th characters.
But the result of the nodesmall mask may contain html tags like .
So the truncation may lead to missing closing html tags or to invalid html tags if the truncation is in the range of the html tag.
Therefor it is better too use css to truncate too long lines.