drop fts table: drop fts table via alembic (#2641)
This commit will change the behaviour of the advanded search
if two or more items with "All data" are selected.

The previous behaviour was that the search for e.g.
All data=A and All data=B only nodes are shown
if A and B is found in the attributes or if A and B is found
in the fulltext.
But no nodes are shown if A is found in the attributes
and B is found in the fulltext or if A is found in the fulltext
and B is found in the attributes.

The new behaviour is that also these nodes are shown
if A is found in the attributes and B is found in the fulltext
or if A is found in the fulltext and B is found in the attributes.

4. commit:
alembic script to drop fts table
fts table is not really dropped but renamed,
to get table back via alembic downgrade