replace `requirements.nix` with package files (#1904)
This commit refactors the way we assemble
the python environemnt for mediatum:

The python build mechanisms are moved to `backend.nix`.
Each python package in `requirements.nix` is moved to
its own files (as the nixpkgs repository does).

Packages are now build with the `callPackage` function that
is provided by nixpkgs python package attrset
(`python.pkgs.callPackage`).
Our packages are added to an extended python package
attrset that is build with the `packageOverrides`
mechanism of the python package attrset.

This has the advanage that new packages are
available to all packages that are now
build with that python package attrset.
However, we have to take care of packages that already exist
in the package attrset provided by nixpkgs:
If we override those, other packages that depend on them
will be rebuild and the total outcome (final deriation) changes.
By renaming those crucial packages (adding version numbers),
we avoid conflicts with upstream packages and ensure
that this commit does not alter the total outcome.

The following python packages override existing ones
(as of NixOS 18.03),
but don't have an effect on other packages we depend on:

* magic
* prettytable
* pytest-catchlog
* selenium
* splinter
* wtforms

The following python packages had to be renamed to
prevent nixpkgs upstream packages from being rebuild:

* attrs -> attrs_18
* imagesize -> imagesize_0_7_1
* pytest -> pytest_2
* snowballstemmer -> snowballstemmer_without_PyStemmer
* sphinx -> sphinx_1_4