add_node_to_xmldoc: detection of exclude_childtypes (#2662)
With #2640 nodes are excluded which type and schema
are in the exclude_childtypes list.
Because the elements of exclude_childtypes list which
comes from the attribute "newnodetype" of the start node of the workflow,
have the form /, the type and schema of the nodes are
concatenated to compare them.
The problem is that some new workflowstep have a none schema,
so the concatenate will fail and this workflowsteps are also excluded.
So the check for nodes in exclude_childtypes must also consider
that a schema may be none.