collections/directories are not displayed completely (#2674)
For search queries of displaying content of collection or
directories a limit of node per page + 1 is used to detect
if there are more entries than nodes per pages.
The problem is that the result for search queries is
automatically distinct but not the result for displaying
collections or directories. So the limit of node per page + 1
produces sometimes results with a lower number of nodes even
if there are more nodes available because sqlalchemy drops
duplicate nodes.
The solution is that the query for content of collection or
directories must be called with distinct().