node_to_file table: remove __versioned__ entry
With the __versioned__ entry in the node_to_file declaration
sqlalchemy-continuum creates a trigger function node_to_file_audit.
This trigger function creates the error duplicate key value
violates unique constraint "node_to_file_version_pkey" if a
file is appended to a new created node.
So this trigger must be removed.
For existing databases the alembic script
73d8e632ec3a_drop_trigger_in_node_to_file.py was used to drop
this trigger.
For new created databases with bin/manage.py schema create
and data init the versioned entry in the node_to_file declaration
must be removed which has the same effect.