web/adminutils: declare default admin menu
There already is a default menu.
However, when showing the menu, the default menu
is needed again to filter certain entries (somehow?).
This very code tries to load the menu from the database,
but *fails* to employ a default value for this lookup.
We declare the default menu globally in the
module file and use it in both places.