show_printview: use path from cte (#2688)
The function getPaths() is very slow (get_accessible_paths()
is normally recommended), but the cte provides also the path
which is used to create the path string.

lru_cache has a default value of 100 for max_size. So it is
better to use 2*len(nodes) for maxsize.
A value of None for max_size for an unlimited cache is a little bit
slower as 2*len(nodes) for maxsize