core/request_handler: fix lrucache for mimetype access (#2809)
The function `core.request_handler.accept_mimetypes`
is equipted with an "lrucache".
However, it accepts the request object as sole argument, which
the cache cannot handle (let alone use as meaningful cache key).

Since the function effectively only uses the "ACCEPT"
argument from the request, we define another function that
accepts the "accept" value and equipt it with the lrucache.
The `accept_mimetypes` function just pulls
\out the "ACCEPT" value and calls the new one.