workflow/addformpage: fix file naming race condition (#2806)
If a form is uploaded into a workflow step,
mediatum tries to find a unique filename on disk.
Sadly, it did so by just counting and checking
if the file exists, not taking into account that
a separate thread/process might just do the same.

This commit replaces the counting mechanism
by just picking a high-entropy random string.
It is simpler and avoids the (security)
risk of the counting race.

Note that the filename on disk
must have a specific format:
It is underscore separated, the final
part will be used as filename when
the file is later emailed to a user.