metadata/upload.py: simplify filetype pattern matching
Using a RegEx here is error prone as
special chars (like a dot '.') in the pattern
are treated as RegEx control chars.
We would have to RegEx-escape the pattern,
but there is no need as the RegEx
merely resembles a `.startwith` method.
Se we just replace it with `.startswith`.

Also, the patterns are the same with and without `fieldname`.