workflow/metafield2metadata.py: new workflowstep (#2885)
This workflow step enables user to assign metadata from metafield to node. Metafield
is stored in workflowstep.attrs.["source-attribute-name"] and specify institute.
Institutes contain (key-value) pairs as metadata are stored in file attached to this
workflow step in form of tree structure.
If processing node.attrs[workflowstep.attrs.["source-attribute-name"]] is found in file
then its attributes will be set or removed to these (key-value) pairs (removed in case
of value is None).