workflow/hierarchicalchoice2metafield.py: new workflowstep (#2885)
This workflowstep step enables user to assign an institute to a document as node in
Database. The hierarchical structure of institutes is stored in file attached to this
workflow step and it is presented to user in form of a selection tree. Fieldname for
institute is stored in workflowstep.attrs["target-attribute-name"]] and selected institute
is stored in node.attrs[workflowstep.attrs["target-attribute-name"]].