get_edit_update_attrs: define to collect update attrs for several nodes (#2827)
When editing multiple nodes in the editor,
the mask calculates the update values for the attributes
of each node -- even though those values are the same for each node.
This commit splits the update method `update_node` into two methods:

* `get_edit_update_attrs`:
  Computes and returns a structure that contains
  all to-be-updated attributes, independent of a node
* `apply_edit_update_attrs_to_node`:
  Uses the precomputed structure to update a single node

As the precomputes needs significant CPU time and Queries
the database for information, this concept increases
the speed of editing multiple nodes at once.