web/edit/edit.py: editing nodes with orphan pathes (#2983)
to show the breadcrumbs to an node to be edited, always the first
parent is selected to get the breadcrumbs until the root node is
reached.
This may lead to a dead end if one of the pathes to the node does
not reach the root node and then an exception is thrown.
Therefore the function get_accessible_paths() is used which only
returns pathes up to the root node.