long list ids in URL: move to data attribute (#3049)
move long ids from URL to data attribute to avoid large request header.
"traditional" is a boolean argument of jquery parameter indicating whether to perform
a traditional "shallow" serialization.
data : { ids : [1,2,3] } will be as "ids=1&ids=2&ids=3" if traditional is set to true,
otherwise the "ids" parameter would have square brackets on the name. Details in
https://api.jquery.com/jQuery.param/, https://stackoverflow.com/questions/31152130/
is-it-good-to-use-jquery-ajax-with-traditional-true/31152304.