schema/schema.py: use children_rel with viewonly (#2940)
Relationships for maskitems and metafields are now defined with
the parameter viewonly=True.
So it is possible to delete metadatatypes, masks, maskitems and
metafields without unlinking them from parent and child nodes.
Without this parameter the parent and child relationsships are
unlinked twice by sqlalchemy, which results in a
StaleDataError: DELETE statement on table 'nodemapping' expected
to delete 1 row(s); Only 0 were matched.