workflow/workflow: move long parameter in URL (#3061)
move long parameter obj as list of node id to data attribute and change request
method to 'POST' to avoid long parameter in URL. Add csrf for POST method.
"traditional" is a boolean argument of jquery parameter indicating whether to perform
a traditional "shallow" serialization.
data : { obj : [1,2,3] } will sent be as "obj=1&obj=2&obj=3" if traditional is set to true,
otherwise the "obj" parameter would have square brackets on the name. Details in
https://api.jquery.com/jQuery.param/, https://stackoverflow.com/questions/31152130/
is-it-good-to-use-jquery-ajax-with-traditional-true/31152304.