core/database/postgres/node: search only in collections (#2933)
Searching all nodes yields a lot of orphan nodes.
Some of those don't have a (valid) scheme assigned,
causing errors when rendered.
To avoid this, we restrict all general searches
to the children of the big Collections node.

This isn't easy as we have to abide by the "offset0"-Trick
to ensure PostgreSQL performance.
The `.search(..)` method no longer retruns a subtree
query object, but instead accepts a filter function argument
that will be used to filter results before `.offset0`
is applied.
Callers no longer have to apply `.offset0`, but instead have
to provide individual filter functions if required.

Two functions that are no longer needed are dropped.