schema.mask.field: maskeditor can't change metatype config (#2940)
When creating or editing a mask item,
one may no longer change the settings
of the underlying meta field.

Caution: The mask item editor also used to modify
the description of the meta field.
This happened through the `{set,get}Description`
method of the mask item class:
It passes on the command to the metafield unless the
mask already has its own description (or has no field).
We change the set-command so it no
longer affects the meta field.
The get-command still needs to "fallback" to the meta field
in order to show nice help boxes to users
(the want to see the meta field description if the mask
item lacks a description, which is the case most of the time).
But in the mask item modification dialog,
the description is directly pulled from the
mask item's attributes; the fallback is *not* used.