web.frontend.content.render_content: `show_id` as return value
Up to this commit, `render_content` returns a parameter
`show_id` in a pass-as-reference fashion:
A list is provided as argument and
filled with one entry by the function.
If no list is provided, the value is not put anywhere.

This is needles cumbersome.
In particular, there is no speed gain by not providing a list.

We drop the list argument.
`render_content` now returns two values:
The html string like before,
and the `show_id` as second return value.
There is only one caller, which is adapted by this commit.

Another function argument is turned from optoinal to positional;
As there is only one caller,
there is no need for optional arguments.