web.frontend.frame.render_page: no `head_meta` without rights
This changes the way `_render_head_meta` is used:
It is applied to the `show_id` or the `node`,
both as they are provided by the caller of `render_page`.
If both are unset, nothing is rendered.

If the user tries to view a node without read access rights,
`show_id` will be `None`, and `node` is also `None`
or represents a container (with read access),
possibly the root collection node.
This is OK, as we have to make sure no metadata of
a unaccessible node is rendered into the meta tags.

The previous implementation (with classical or TAL masks)
permitted access to head_meta of nominally blocked nodes:

  ``/?id=1&show_id=2

rendered the metadata of node 2
irrespective of its access rights.

The commit at hand ensures the either the
target node (with access rights) or no
node at all gets its meta tags rendered.

Thanks to M. Grummann for pointing this out
and to N. Kolbenschlag for investigating!