updateMetaField: assert bool(fieldid)!=bool(fieldtype) (#2940)
This function processes two cases:

* new field: field type `mtype` is defined, but the field
  does not exist yet, so `fieldid` must be undefined
* edit field: field id `fieldid` is defined, but the field
  type may not be changed for existing fields,
  so `mtype` must be undefined

The corresponding TAL macros are modified:
The value that must be undefined is no
longer passed back to the request handler.

This implies another change, however:
For existing fields, the request handler must extract
the meta field configuration `{type}_value` from the
request arguments and therefore needs the field type.
Since `mtype` is no longer available, we use the `fieldid`
to obtain the meta field object and get its meta type.

We also drop the "search field by name" code as it is not used.