Metatype.formatValues: dissolve `context`, yield values (#2940)
Lists define a method `formatValues`.
This commit replaces the single `context` argument with
three arguments that contain all the information needed.
The function itself is now a generator
and yields its result as 4-Tuples.

Besides, some callers compare the
resulting tuples with an empty string.
This makes no sense as this comparison always returns `False`,
so the comparison (and the filter that employs it) is dropped.