web.edit.edit.content: move search call down (read-only) (#3128)
Searching for nodes sets the session read-only.
We must avoid writing to the database afterwards.

The function `handletabs` updates some sort settings in the
database, so it must be called *before* any searches are
performed, which would turn the DB session into read-only mode.

The general call `_editModules[current].getContent(req, ids)`
*can* perform a search (depending on the `current` module).
So it must happen *after* `handletabs`.

`handletabs` should (probably?) not
be called if `values["style"]=="popup"`.
We cannot just move the check-and-return block
for this above the `handletabs` call because
`_editModules[current].getContent(req, ids)`
must be called *in any case*.
So we enclose the call to `handletabs`
in another if-block that checks for the "style" value.