metadata/htmlmemo.py: add parameter wysiwyg (#3143)
With the parameter wysiwyg it can be distinguised if a htmlmemo
field can be edited as a htmlmemo field (old behaviour) or a
memo field (wysiwyg is false) for a future remove of the memo
metatype.
An alembic script adds the parameter wysiwyg with to value True
to all existing htmlmemo fields.