new editor module versioning (#3135)
The node versioning feature can no longer be accessed
via the metadata editor or the file upload function.
This resulted in a confusing user interface as it wasn't
clear if the old or new version got persisted.
Also, the old interface was able to handle more than one
node at a time, leading to inconsistencies in the database
if two nodes with different version count are edited at once.

The new interface is a separate versioning module and dialog
which lists all the given versions for a node and
permits to persist a new version, based on the current version.
The new module only handles one node at a
time to avoid database inconsistencies.

The new module is reachable via the editor menu.
As we don't have a nice icon at hand,
we just use the letter "V" as icon for now.