core/translation.py: restructure translate function (#3063)
key to be translated can be not defined in po files. Split translate into
translate and translate_in_request. Language is available as parameter in
translate and language will be taken from flask request if language is not
given. Define translate_in_template that handles key error to use in template.