web.frontend.main._handle_json_request: clearly split classes (#3139)
This function used to call the method `getSearchHTML` of a
field-like object, without known which class it was using.
We want to rename some methods of Metatype classes,
which would break this function as it no
longer knows which method name to use.
So this commit adds a clear disticntion between the
possible classes which will later help renaming mehtods.

Note that the `field` parameter for `getSearchHTML`
was only given in the Metatype case.
However, in that case, the `field` parameter is
never used (by the text Metatype, at least).
So we can fix this parameter to `None`.